refactor(rename): rename mod: `fzf_make` to `tui`, `fzf_make_main` to `fzf_make`
