+++ src/usecase/mod.rs
@@ -1,9 +1,9 @@
 mod execute_make_command;
 pub(super) mod fzf_make;
-pub(super) mod fzf_make_main;
 pub(super) mod help;
 pub(super) mod history;
 pub(super) mod invalid_arg;
 pub(super) mod repeat;
+mod tui;
 pub(super) mod usecase_main;
 pub(super) mod version;
